Moved DartLocalization.py to root difrectory
